requirements.txt
langchain
langchain-google-genai
google-generativeai
streamlit
python-dotenv
PyPDF2
langchain_community
langchain_huggingface 

-e .